task8/requirements.txt
pandas~=1.5.1
boto3~=1.26.3
botocore~=1.29.3
apache-airflow